dev-simbricks: add SimBricks PCI adapter
PCI device adapter for the SimBricks simulation framework. SimBricks
connects multiple simulators of different types together, including time
synchronization. This adapter connects a SimBricks PCI simulator as a
PCI device into gem5. The primary use-case, at the moment, is for
network cards, but the adapter is independent of this. Supports MSI &
MSI-X.

Change-Id: I38e8c4311886a35a46972448fd2c8685ed5574e1