dev: allow PIO devices to override PIO port impl.
Instead of directly accessing the protected pioPort member, we add a
virtual function getPioPort() that returns the port to use. Also
change users to refer to this.

Motivating use case for this is the SimBricks PCI adapter where we need
to be able to do asynchronous timing-protocol memory accesses. We expect
there will be other use-cases for this as well (with things like CXL
coming etc.).

Change-Id: Idd7148357b60d4b6387260c2966d55f839d2ee45